requirements.txt
blast==2.12.0
python==3.8.18
biom-format=2.1.12
biopython=1.79
epa-ng=0.3.8
gappa=0.8.0
hmmer=3.3.2
joblib=1.1.0
numpy=1.22.3
pandas=1.4.2
r-castor=1.7.11